requirements_test.txt
coveralls
pylint
pytest
pytest-cov
pytest-django

